requirements.txt
streamlit==1.31.0
openai==1.40.0
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.0
pydantic==2.8.2
